pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "journeylogger"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "python-dotenv==1.1.0",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.2.2",
    "gspread==6.2.1",
    "httplib2==0.22.0",
    "idna==3.10",
    "oauth2client==4.1.3",
    "oauthlib==3.2.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyparsing==3.2.3",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "python-telegram-bot",
    "urllib3==2.4.0",
    "pytest==8.4.0",
]

[project.scripts]
journeylogger = "journeylogger.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]